Trigger openURL logic when build ID changes. (#403)
In Firefox, the openURL action on update is only triggered if the
milestone (i.e. version number) changes. The milestone corresponds to
the platform version, which is the firefox version we're using. That
means that a milestone update was not triggered for us if we updated our
version number, but were still using the same firefox version in the
build.

This changes the behaviour to trigger openURL for both milestone changes
and buildID changes, so we should be able to trigger the openURL action
on every update.