Use accaptable version name
On Windows builds .res files for the main executable - Ghostery.exe.res
are produced with invalid value of PRODUCTVERSION.

Code that generates the version is here https://github.com/mozilla/gecko-dev/blob/aa46c2dcccbc6fd4265edca05d3d00cccdfc97b9/config/create_rc.py#L133

Without getting into much details, version name 2021.07 is correct but 2021.08 or 2021.09 is not. This is because of "07" for .rc file (resource file / C++) is a valid number but "08" and "09" are not.

This issue was generating build crash with:
25:05.04 browser/app/Ghostery.exe.res
25:05.04 llvm-rc: Error parsing file: Integer invalid or too large: 09